Migrated all functions accessing directly to the database `users` table to database_users_table_util.
